Corrections to work with command-line arguments. "home" command. Some extra exception catching. Changed "rename" command logic.

+++ src/localhost/Main.java
@@ -7,43 +7,52 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.Scanner;
 
-import static localhost.Checks.*;
+import static localhost.Checks.correctArg;
+import static localhost.Checks.pathExists;
 
 public class Main {
 
     public static void main(String[] args) throws IOException {
         Scanner scanner = new Scanner(System.in);
-        String input;
         Target target = new Target();
         Target temp = new Target();
         Target pick = new Target();
-        System.out.println(Arrays.deepToString(args));
+        List<String> spell;
+        int sSize;
 
         progcycle:
         while (true) {
             System.out.println("\nTarget: " + target.get().toString());
             System.out.println("Picked: " + pick.get().getFileName().toString() + "\n");
-            input = scanner.nextLine().trim();
-            List<String> spell = new ArrayList<>(Utils.findArgs(input.split(" ")));
-            int sSize = spell.size();
+
+            if (args.length == 0) {
+                args = scanner.nextLine().trim().split(" ");
+                spell = Utils.findArgs(args);
+            } else {
+                spell = new ArrayList<>(Arrays.asList(args));
+                spell.add(null);
+            }
+            args = new String[]{};
+            sSize = spell.size();
 
             spellcycle:
-            for (int word = 0; word < sSize; word++) {
-                String lastWord = (word == 0 ? null : spell.get(word - 1));
-                String nextWord = (word == sSize - 1 ? null : spell.get(word + 1));
+            for (int word = 0; word < sSize - 1; word++) {
+                String nextWord = spell.get(word + 1);
                 String thisWord = spell.get(word);
-
                 switch (thisWord.toLowerCase()) {
+                    case "home":
+                        target.set(Paths.get(System.getProperty("user.dir")));
+                        break;
                     case "size":
                         System.out.println("Size of " + target.get().getFileName() + " is " + Files.size(target.get()) + " bytes.");
                         break;
                     case "show":
-                    case "contents":
+                    case "list":
+                    case "dir":
+                    case "look":
                         try (DirectoryStream<Path> stream = Files.newDirectoryStream(target.getDir())) {
-                            int i = 1;
                             for (Path file : stream) {
-                                System.out.print("\t" + file.getFileName() + "\t");
-                                if(i++ % 4 == 0) System.out.print("\n");
+                                System.out.println("\t" + file.getFileName());
                             }
                         }
                         break;
@@ -61,8 +70,11 @@ public class Main {
                         pick.set(pick.last());
                         break;
                     case "log":
-                        if (wordExists(nextWord) && Files.isRegularFile(target.get())) {
-                            Files.write(target.get(), nextWord.substring(1).getBytes(), StandardOpenOption.CREATE, StandardOpenOption.APPEND);
+                        if (correctArg(spell.get(word + 1))) {
+                            if (Files.isRegularFile(target.get())) {
+                                String s = nextWord + System.getProperty("line.separator");
+                                Files.write(target.get(), s.getBytes(), StandardOpenOption.CREATE, StandardOpenOption.APPEND);
+                            } else System.out.println("Your target is not a file!");
                         }
                         word++;
                         break;
@@ -73,6 +85,7 @@ public class Main {
                         break;
                     case "copy":
                     case "paste":
+                    case "clone":
                         temp.set(target.getDir().resolve(pick.get().getFileName()));
                         if (Files.exists(temp.get())) {
                             System.out.println(temp.get().toString() + " is already exists!");
@@ -80,10 +93,17 @@ public class Main {
                             Files.copy(pick.get(), temp.get(), StandardCopyOption.COPY_ATTRIBUTES);
                             target.set(temp.get());
                         }
+                        word++;
                         break;
                     case "move":
                     case "replace":
-                        temp.set(target.getDir().resolve(pick.get().getFileName()));
+                        try {
+                            temp.set(target.getDir().resolve(pick.get().getFileName()));
+                        } catch (InvalidPathException e) {
+                            System.out.println("You used invalid characters!");
+                            continue;
+                        }
+
                         if (Files.exists(temp.get())) {
                             System.out.println(temp.get().toString() + " is already exists!");
                         } else {
@@ -93,24 +113,34 @@ public class Main {
                         }
                         break;
                     case "rename":
-                        temp.set(pick.getDir().resolve(nextWord.substring(1)));
-                        if (Files.exists(temp.get())){
-                            System.out.println(temp.get().toString() + " is already exists!");
-                        } else {
-                            Files.move(pick.get(), temp.get());
-                            Files.deleteIfExists(pick.get());
-                            target.set(temp.get());
+                        if (correctArg(nextWord)) {
+                            try {
+                                temp.set(pick.get().getParent().resolve(nextWord));
+                            } catch (InvalidPathException e) {
+                                System.out.println("You used invalid characters!");
+                                continue;
+                            }
+                            if (Files.exists(temp.get())) {
+                                System.out.println(temp.get().toString() + " is already exists!");
+                            } else {
+                                Files.move(pick.get(), temp.get());
+                                Files.deleteIfExists(pick.get());
+                                target.set(temp.get());
+                            }
                         }
+                        word++;
                         break;
                     case "up":
-                        target.set(target.get().getParent());
+                    case "..":
+                    case "parent":
+                        target.set(target.getDir().getParent());
                         break;
                     case "new":
                     case "create":
                     case "make":
                         if (correctArg(nextWord)) {
                             try {
-                                temp.set(target.getDir().resolve(nextWord.substring(1)));
+                                temp.set(target.getDir().resolve(nextWord));
                             } catch (InvalidPathException e) {
                                 System.out.println("You used invalid characters!");
                                 continue;
@@ -131,9 +161,16 @@ public class Main {
                     case "root":
                         target.set(target.get().getRoot().toAbsolutePath());
                         break;
+                    case "go":
                     case "goto":
-                        if (correctArg(nextWord)) {
-                            temp.set(nextWord);
+                    case "path":
+                        if (correctArg(spell.get(word + 1))) {
+                            try {
+                                temp.set(nextWord);
+                            } catch (InvalidPathException e) {
+                                System.out.println("You used invalid characters!");
+                                continue;
+                            }
                             if (pathExists(temp.get())) {
                                 target.set(temp.get());
                             }
@@ -146,15 +183,18 @@ public class Main {
                     case "stop":
                         break spellcycle;
                     default:
-                        if (thisWord.charAt(0) == '|') {
-                            temp.set(target.getDir().resolve(thisWord.substring(1)));
-                            if (pathExists(temp.get())) {
-                                target.set(temp.get());
-                            }
-                        } else {
-                            System.out.println("\"" + thisWord + "\" is not correct magic word.");
+                        if (thisWord.charAt(0) == '\\' || thisWord.charAt(0) == '/') {
+                            thisWord = thisWord.substring(1);
                         }
-                        if (sSize <= 1) break spellcycle;
+                        try {
+                            temp.set(target.getDir().resolve(thisWord));
+                        } catch (InvalidPathException e) {
+                            System.out.println("You used invalid characters!");
+                            continue;
+                        }
+                        if (pathExists(temp.get())) {
+                            target.set(temp.get());
+                        }
                 }
             }
         }
